+++ ghost/core/package.json
@@ -100,7 +100,6 @@
     "@tryghost/limit-service": "1.2.15",
     "@tryghost/link-replacer": "0.0.0",
     "@tryghost/logging": "2.4.21",
-    "@tryghost/member-attribution": "0.0.0",
     "@tryghost/members-csv": "0.0.0",
     "@tryghost/members-offers": "0.0.0",
     "@tryghost/metrics": "1.0.37",
@@ -113,6 +112,7 @@
     "@tryghost/pretty-cli": "1.2.46",
     "@tryghost/prometheus-metrics": "0.0.0",
     "@tryghost/promise": "0.3.14",
+    "@tryghost/referrer-parser": "0.1.1",
     "@tryghost/request": "1.0.10",
     "@tryghost/root-utils": "0.3.32",
     "@tryghost/security": "0.0.0",
