+++ ghost/core/test/unit/server/services/member-attribution/history.test.js
@@ -1,6 +1,6 @@
-require('should');
+const should = require('should');
 
-const UrlHistory = require('../lib/UrlHistory');
+const UrlHistory = require('../../../../../core/server/services/member-attribution/UrlHistory');
 
 describe('UrlHistory', function () {
     it('sets history to empty array if invalid', function () {
