--- ghost/member-attribution/index.js
@@ -1,7 +1,0 @@
-module.exports = {
-    MemberAttributionService: require('./lib/MemberAttributionService'),
-    AttributionBuilder: require('./lib/AttributionBuilder'),
-    UrlTranslator: require('./lib/UrlTranslator'),
-    ReferrerTranslator: require('./lib/ReferrerTranslator'),
-    OutboundLinkTagger: require('./lib/OutboundLinkTagger')
-};
